telegram-bot/requirements.txt
# Telegram Bot Requirements
# Estif Bingo Telegram Bot

# Core Dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0

# Database
asyncpg==0.29.0

# Security
pyjwt==2.8.0
bcrypt==4.1.0
cryptography==41.0.7

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
urllib3==2.1.0

# Optional - for payment webhooks (if needed)
# stripe==6.5.0
# chapa==1.0.0

# Development dependencies (optional)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0